+++ webui/react/src/components/JupyterLabGlobal.tsx
@@ -2,11 +2,11 @@ import React, { useEffect } from 'react';
 
 import JupyterLabModalComponent from 'components/JupyterLabModal';
 import { useModal } from 'components/kit/Modal';
-import { keyEmitter, KeyEvent } from 'hooks/useKeyTracker';
-import { useSettings } from 'hooks/useSettings';
 import shortCutSettingsConfig, {
   Settings as ShortcutSettings,
-} from 'pages/Admin/UserSettings.settings';
+} from 'components/UserSettings.settings';
+import { keyEmitter, KeyEvent } from 'hooks/useKeyTracker';
+import { useSettings } from 'hooks/useSettings';
 import { Workspace } from 'types';
 import { matchesShortcut } from 'utils/shortcut';
